+++ src/api/v1/auth.py
@@ -1,9 +1,6 @@
 import logging
-from base64 import b64encode
 from typing import Annotated
-from urllib.parse import urlencode
 
-import requests
 from fastapi import (
     APIRouter,
     Body,
@@ -17,7 +14,6 @@ from fastapi import (
 )
 from pydantic import BaseModel
 
-from core.config import settings
 from responses.auth_responses import (
     get_change_psw_response,
     get_login_response,
@@ -26,7 +22,6 @@ from responses.auth_responses import (
 )
 from schemas.auth import UserLogin, UserLoginResponse, UserTokenResponse
 from schemas.user import UserCreate, UserRead, UserUpdate
-from schemas.yndx_oauth import UserInfoSchema
 from services.auth.auth_service import AuthService, get_auth_service
 from services.utils import (
     get_user_id_from_access_token,
@@ -100,83 +95,6 @@ async def login_user(
         httponly=True,
         samesite="lax",
     )
-    return user
-
-
-@router.get(
-    "/yndx_social_login",
-    status_code=status.HTTP_200_OK,
-    summary="OAuth 2.0 Yndx login",
-    description="Redirects to Yandex OAuth 2.0 login page.",
-)
-async def yndx_social_login():
-    params = {
-        "response_type": "code",
-        "client_id": settings.yndx_oauth.YNDX_CLIENT_ID,
-    }
-    auth_url = f"{settings.yndx_oauth.YNDX_CODE_URL}?{urlencode(params)}"
-    logger.warning("url: %s", auth_url)
-    return {"url_for_auth": auth_url}
-
-
-@router.get(
-    "/yndx_callback",
-    response_model="",
-    status_code=status.HTTP_200_OK,
-    summary="OAuth 2.0 Yndx login",
-    description="Redirects to Yandex OAuth 2.0 login page.",
-)
-async def yndx_callback(
-    response: Response,
-    code: str = Query(
-        ...,
-        title="code",
-        description="Код подтверждения аутентификации в Yandex",
-    ),
-    auth_service: AuthService = Depends(get_auth_service),
-    user_agent: Annotated[str | None, Header()] = None,
-):
-    client_id = settings.yndx_oauth.YNDX_CLIENT_ID
-    client_secret = settings.yndx_oauth.YNDX_CLIENT_SECRET
-    raw_str = f"{client_id}:{client_secret}"
-    encoded_creds = b64encode(raw_str.encode()).decode()
-
-    data = {"grant_type": "authorization_code", "code": code}
-
-    headers = {
-        "Content-type": "application/x-www-form-urlencoded",
-        "Authorization": f"Basic {encoded_creds}",
-    }
-
-    token_resp = requests.post(
-        url=settings.yndx_oauth.YNDX_TOKEN_URL, headers=headers, data=data
-    )
-    resp_dict = token_resp.json()
-    logger.warning("Response YNDX_TOKEN_URL: %s", resp_dict)
-    headers = {
-        "Authorization": f"OAuth {resp_dict['access_token']}",
-    }
-
-    response_info = requests.get(
-        url=settings.yndx_oauth.YNDX_INFO_URL, headers=headers
-    )
-    logger.warning("REsponse: %s", response_info.json())
-    user_info = UserInfoSchema(**response_info.json())
-    user, tokens = await auth_service.login_user_yndx(user_info, user_agent)
-
-    response.set_cookie(
-        key="access_token",
-        value=tokens.access_token,
-        httponly=True,
-        samesite="lax",
-    )
-    response.set_cookie(
-        key="refresh_token",
-        value=tokens.refresh_token,
-        httponly=True,
-        samesite="lax",
-    )
-
     return user
